+++ Instructor/admin.py
@@ -1,0 +1,1 @@
+# Register your models here.
