+++ Instructor/templates/rubrics/rubric_delete.html
@@ -1,10 +1,10 @@
-{% extends 'base.html' %}
+{% extends 'admin_base.html' %}
 
 {% block mainContent %}
     <form method="post">
         {% csrf_token %}
         <h1>Are you sure you want to delete rubric: <b>{{ rubric.name }}?</b></h1>
         <button>Yes</button>
-        <a href="{% url "rubric_list" %}">Cancel</a>
+        <a href="{% url "rubric-list" %}">Cancel</a>
     </form>
 {% endblock %}
