+++ Main/admin.py
@@ -1,9 +1,8 @@
-from django.contrib import admin
 from django.conf import settings
+from django.contrib import admin
 
 from . import models
 
-
-if settings.DEBUG:
-    admin.site.register(models.Rubric)
-    admin.site.register(models.Review)
+# if settings.DEBUG:
+#     admin.site.register(models.Rubric)
+#     admin.site.register(models.Review)
