+++ Main/static/js/rubrics/rubric-grade-widget.js
@@ -1,15 +1,15 @@
 'use strict';
 
 function update_cell_colors() {
-    $(".rubric-grading td.cell").each(function(index, object) {
+    $(".rubric-grading td.cell").each(function (index, object) {
         console.log($(object).children("input:checked").length);
         $(object).toggleClass("selected", $(object).children("label").children("input:checked").length > 0)
     });
 }
 
-function get_scores(){
+function get_scores() {
     let scores = [];
-    $(".rubric-grading .criteria input:checked").each(function(index, object) {
+    $(".rubric-grading .criteria input:checked").each(function (index, object) {
         scores.push(parseFloat($(object).val()));
     });
     return scores;
@@ -25,13 +25,14 @@ function get_max(row_index) {
 
 function update_total_score(scores) {
     const report = $("#grade_report");
-    let report_lst = report.text().split("/")
+    let report_lst = []
     let total = 0;
     let max = 0;
-    scores.forEach(score => {
-        total += score === -1? 0 : score;
-        max += score === -1? 0 : get_max(scores.indexOf(score));
-    });
+    for (let i = 0; i < scores.length; i++) {
+        let score = scores[i];
+        total += score === -1 ? 0 : score;
+        max += score === -1 ? 0 : get_max(i);
+    }
     report_lst[0] = `Grade: ${total.toFixed(1)}`;
     report_lst[1] = max.toFixed(1);
     report.text(report_lst.join('/'));
@@ -44,13 +45,16 @@ function score_button_callback() {
     update_total_score(scores);
 }
 
-function score_cell_callback(event){
+function score_cell_callback(event) {
     $(event.target).children('label').children(".score-select").click();
 }
 
-function  loadCallback(event){
+function loadCallback() {
     $(".score-select").click(score_button_callback);
     $(".cell").click(score_cell_callback);
+    $(".score-select:checked").each((index, object) => {
+        $(object).click();
+    });
 }
 
 $(document).ready(loadCallback);
