+++ Main/templates/base.html
@@ -17,13 +17,17 @@
 {#  Navigation goes here #}
 {# For now it'll be a list of links #}
 <a aria-label="Home" href="{% url "home" %}"><h1>Code Review</h1></a>
-<div style="display: flex; flex-direction: row; float: right;">
-    <a style="margin-right: 30px;" href="{% url "logout" %}">Logout</a>
-    <a style="margin-right: 30px;" href="{% url "review_create" %}">New Review</a>
-    {% if user.is_superuser %}
-        <a style="margin-right: 30px;" href="{% url "rubric_list"%}">Rubrics</a>
-        <a href="/admin">Admin</a>
-    {% endif %}
+<div class="nav-div">
+    {% block nav %}
+        <a href="{% url "review-create" %}">New Review</a>
+        {% if user.is_superuser %}
+            <a href="{% url "instructor-home" %}">Admin</a>
+            {% if debug %}
+                <a href="/admin">Debug</a>
+            {% endif %}
+        {% endif %}
+        <a href="{% url "logout" %}">Logout</a>
+    {% endblock %}
 </div>
 {% block mainContent %}
