+++ Main/templates/reviews/review_cancel.html
@@ -3,7 +3,8 @@
 {% block mainContent %}
     <form method="post">
         {% csrf_token %}
-        <h1>Are you sure you want to {% block verb %}cancel{% endblock %} review with Schoology id: <b>{{ review.schoology_id }}</b>?</h1>
+        <h1>Are you sure you want to {% block verb %}cancel{% endblock %} review with Schoology id:
+            <b>{{ review.schoology_id }}</b>?</h1>
         <button>Yes</button>
         <a href="{% url "home" %}">Cancel</a>
     </form>
