+++ Main/templates/reviews/review_delete.html
@@ -1,0 +1,3 @@
+{% extends "reviews/review_cancel.html" %}
+
+{% block verb %}delete{% endblock %}
