--- Main/templates/rubrics/rubric_list.html
@@ -1,18 +1,0 @@
-{% extends 'base.html' %}
-
-{% block mainContent %}
-    <table>
-        <tbody>
-            {% for rubric in rubrics %}
-                <tr>
-                    <td>{{ rubric.name }}</td>
-                    <td><a href="{% url "rubric_edit" rubric.id %}">Edit</a></td>
-                    <td><a href="{% url "rubric_del" rubric.id %}">Delete</a></td>
-                </tr>
-            {% endfor %}
-            <tr>
-                <td colspan="2"><a href="{% url "rubric_create" %}">Add New Rubric</a></td>
-            </tr>
-        </tbody>
-    </table>
-{% endblock %}
