+++ Main/templates/widgets/rubric_grade.html
@@ -8,11 +8,15 @@
     </tr>
     {% for row in widget.rubric|rows %}
         <tr id="grade-row-{{ row.index }}" class="criteria">
-            <th>{{ row.name }} <br> {{ row.description }} <br> Max: <span class="max-score-val">{{ row.max_score }}</span></th>
+            <th>{{ row.name }} <br> {{ row.description }} <br> Max: <span
+                    class="max-score-val">{{ row.max_score }}</span></th>
             {% for cell in row|cells %}
-                <td class="cell"> <span class="score-val">{{ cell.score }}</span> <br> <label>{{ cell.description }} <br> <input class="score-select" value="{{ cell.score }}" name="{{ row.name }}_score" type="radio"/></label> </td>
+                <td class="cell"><span class="score-val">{{ cell.score }}</span> <br> <label>{{ cell.description }} <br>
+                    <input class="score-select" value="{{ cell.score }}" name="{{ row.name }}_score"
+                           type="radio"/></label></td>
             {% endfor %}
-            <td class="cell"> <span class="score-val">N/A</span> <br> <label>N/A <br> <input class="score-select" value="-1" name="{{ row.name }}_score" type="radio"/></label> </td>
+            <td class="cell"><span class="score-val">N/A</span> <br> <label>Does Not Apply <br> <input
+                    class="score-select" value="-1" name="{{ row.name }}_score" type="radio"/></label></td>
         </tr>
     {% endfor %}
 </table>
